refactor(93): Add worker count in build info
