+++ src/main/kotlin/io/github/janbarari/gradle/analytics/scanner/execution/BuildExecutionLogicImp.kt
@@ -82,6 +82,7 @@ class BuildExecutionLogicImp(
     private val tower: Tower,
     private val requestedTasks: List<String>,
     private val modules: Set<Module>,
+    private val gradleWorker: Pair<Int, Int>,
     private val saveMetricUseCase: SaveMetricUseCase,
     private val saveTemporaryMetricUseCase: SaveTemporaryMetricUseCase,
     private val upsertModulesTimelineUseCase: UpsertModulesTimelineUseCase,
@@ -170,6 +171,8 @@ class BuildExecutionLogicImp(
                 )
             )
 
+        buildMetric.gradleWorkers = gradleWorker
+
         saveTemporaryMetricUseCase.execute(buildMetric)
         saveMetricUseCase.execute(buildMetric)
 
@@ -206,6 +209,10 @@ class BuildExecutionLogicImp(
             printLine(left = "Overall Build Process:", right = "${buildMetric.overallBuildProcessMetric?.median}ms")
             printLine(left = "Cache Hit:", right = "${buildMetric.cacheHitMetric?.rate}%")
             printLine(left = "Parallel Execution Rate:", right = "${buildMetric.parallelExecutionRateMetric?.medianRate}%")
+            printLine(
+                left = "Used/Available Worker Count:",
+                right = "${buildMetric.gradleWorkers.first}/${buildMetric.gradleWorkers.second}"
+            )
             printBreakLine(char = '-')
             printLine(left = "Datetime:", right = DateTimeUtils.formatToDateTime(buildMetric.createdAt))
             printBreakLine(char = '-')
