+++ src/test/kotlin/io/github/janbarari/gradle/MockSetProperty.kt
@@ -23,21 +23,20 @@
 package io.github.janbarari.gradle
 
 import org.gradle.api.Transformer
-import org.gradle.api.provider.ListProperty
-import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
+import org.gradle.api.provider.SetProperty
 import org.gradle.internal.impldep.org.apache.http.MethodNotSupportedException
 import java.util.function.BiFunction
 
-class ListPropertyMock<T>(var value: MutableList<T>?): ListProperty<T> {
-    override fun get(): MutableList<T> = value!!
+class MockSetProperty<T>(var value: MutableSet<T>?): SetProperty<T> {
+    override fun get(): MutableSet<T> = value!!
 
-    override fun getOrNull(): MutableList<T>? = value
+    override fun getOrNull(): MutableSet<T>? = value
 
     override fun isPresent(): Boolean = value != null
 
     @Deprecated("", ReplaceWith("this"))
-    override fun forUseAtConfigurationTime(): Provider<MutableList<T>> {
+    override fun forUseAtConfigurationTime(): Provider<MutableSet<T>> {
         return this
     }
 
@@ -57,20 +56,20 @@ class ListPropertyMock<T>(var value: MutableList<T>?): ListProperty<T> {
         // do nothing
     }
 
-    override fun empty(): ListProperty<T> {
+    override fun empty(): SetProperty<T> {
         return this
     }
 
-    override fun convention(provider: Provider<out MutableIterable<T>>): ListProperty<T> {
+    override fun convention(provider: Provider<out MutableIterable<T>>): SetProperty<T> {
         return this
     }
 
-    override fun convention(elements: MutableIterable<T>?): ListProperty<T> {
+    override fun convention(elements: MutableIterable<T>?): SetProperty<T> {
         return this
     }
 
     override fun addAll(provider: Provider<out MutableIterable<T>>) {
-        throw MethodNotSupportedException("addAll(provider) is not supported in the ListPropertyMock")
+        throw MethodNotSupportedException("addAll(provider) is not supported in the SetPropertyMock")
     }
 
     override fun addAll(elements: MutableIterable<T>) {
@@ -89,46 +88,46 @@ class ListPropertyMock<T>(var value: MutableList<T>?): ListProperty<T> {
         value?.add(element)
     }
 
-    override fun value(provider: Provider<out MutableIterable<T>>): ListProperty<T> {
+    override fun value(provider: Provider<out MutableIterable<T>>): SetProperty<T> {
         return this
     }
 
-    override fun value(elements: MutableIterable<T>?): ListProperty<T> {
+    override fun value(elements: MutableIterable<T>?): SetProperty<T> {
         return this
     }
 
     override fun set(provider: Provider<out MutableIterable<T>>) {
-        throw MethodNotSupportedException("set(provider) is not supported in the ListPropertyMock")
+        throw MethodNotSupportedException("set(provider) is not supported in the SetPropertyMock")
     }
 
     override fun set(elements: MutableIterable<T>?) {
-        this.value = elements?.toMutableList()
+        this.value = elements?.toMutableSet()
     }
 
     override fun <U : Any?, R : Any?> zip(
         p0: Provider<U>,
-        p1: BiFunction<in MutableList<T>, in U, out R>
+        p1: BiFunction<in MutableSet<T>, in U, out R>
     ): Provider<R> {
-        throw MethodNotSupportedException("zip is not supported in the ListPropertyMock")
+        throw MethodNotSupportedException("zip is not supported in the SetPropertyMock")
     }
 
-    override fun orElse(p0: Provider<out MutableList<T>>): Provider<MutableList<T>> {
+    override fun orElse(p0: Provider<out MutableSet<T>>): Provider<MutableSet<T>> {
         return this
     }
 
-    override fun orElse(value: MutableList<T>): Provider<MutableList<T>> {
+    override fun orElse(value: MutableSet<T>): Provider<MutableSet<T>> {
         return this
     }
 
-    override fun <S : Any?> flatMap(transformer: Transformer<out Provider<out S>, in MutableList<T>>): Provider<S> {
-        throw MethodNotSupportedException("flatMap is not supported in the ListPropertyMock")
+    override fun <S : Any?> flatMap(transformer: Transformer<out Provider<out S>, in MutableSet<T>>): Provider<S> {
+        throw MethodNotSupportedException("flatMap is not supported in the SetPropertyMock")
     }
 
-    override fun <S : Any?> map(transformer: Transformer<out S, in MutableList<T>>): Provider<S> {
-        throw MethodNotSupportedException("map is not supported in the ListPropertyMock")
+    override fun <S : Any?> map(transformer: Transformer<out S, in MutableSet<T>>): Provider<S> {
+        throw MethodNotSupportedException("map is not supported in the SetPropertyMock")
     }
 
-    override fun getOrElse(defaultValue: MutableList<T>): MutableList<T> {
+    override fun getOrElse(defaultValue: MutableSet<T>): MutableSet<T> {
         return value!!
     }
